feat(rslint_parser): Replace `Syntax` with `SourceType` (#2159)
The definition of the existing `Syntax` struct isn't flexible enough to fully support the different TS/JS combinations. For example:

* TypeScript files can be modules or scripts.
* TypeScript supports declaration files
* The support of top level await depends on the language version

This PR replaces the `Syntax` struct with a `SourceType` struct that consist of:

```
pub struct SourceType {
    language: Language,
    variant: LanguageVariant,
    module_kind: ModuleKind,
    version: LanguageVersion,
}
```

* Language: JS or TS (TS has a flag for declaration file)
* variant: Standard or JSX
* module_kind: Module / Script (maybe CommonJS at some point?)
* version: Not used today but the ECMA Script version of the file

Correctly representing this combination revealed that:

* TypeScript files weren't parsed as a module (or in strict mode)
* A TS file can have multiple `export default` for function overload or if they appear in different ambient contexts (e.g. two `declare module` blocks).

The current implementation infers the `SourceType` from the file extension. This won't be enough in the long run. TypeScript uses the `tsconfig.json` to determine if files should be parsed as Script or Module.

This struct isn't intended to be used to determine which features should be transpiled during a transform. I believe that we'll want a more granular representation so that individual syntax can be toggled on and off.