fix(rslint_parser): Fix ASI for return type `is`/`assert` predicates
TypeScript supports `: <ident> is <condition>` and `asserts <ident>` predicates on function return types.

This PR ensures that our Parser correctly respects ASI if there's a line break right before the `is` or `asserts` token:

```ts
interface I {
    foo(callback: (a: any, b: any) => void): I
    is(): boolean;
    bar(test: string): I
    asserts(): boolean
}
```

`is` and `asserts` are both methods in this example and shouldn't be parsed as return type predicates (`I is...` or `I asserts...`).