chore(coverage): assert no unknown in coverage (#2171)
Extends the coverage tests to assert that the parsed result contains no `Unknown` nodes for tests that are expected to pass. 

This revealed that our Unicode escape sequence parsing is too strict. The validation if Unicode escape sequences are valid (form a valid character) is only necessary in export-name positions (see T262 regressions) but not any others. This check was added in #1938 and is reverted in this PR because I believe it's more important that we parse valid files correctly. We can correctly implement the check once the work on #2035 is done. 

This PR further fixes the parsing of `getter` object members to parse out a normal type annotation and not a return type annotation.