fix(rslint_parser): Model class member modifiers as array
TypeScript doesn't enforce a strict class member modifier ordering. For example, all the following
modifier orders are valid:

```ts
class Test {
    private declare c?: string;
    declare private d?: string;
    declare private readonly e: string;
    private readonly declare f: string;
    protected abstract readonly g: string;
    protected readonly abstract h: string;
}
```

`declare` can be the first modifier or come last, same for `abstract`.

This PR replaces the modifier tokens on the individual class members with lists, so that different modifier orders can be supported. I decided to use different lists for the cases where a member supports different modifiers so that an analyzer doesn't need to handle modifiers that can never be present for a specific node.j

This comes at the cost that it's no longer possible to support refactors by enforcing the correct order.