chore(coverage): Infer should error for ts/babel tests (#2193)
Not all tests in the `typescript` directory are expected to pass. A test is expected to fail if the `options.json` has a `throws` property or the `output.json` contains any `SyntaxError`

This PR infers the expected outcome by matching the content of the `options.json` and `output.json` (parsing the JSON feels overkill).

There are still plenty of tests that should pass but fail. However, these tests do contain syntax errors. I'm not sure if it's worth keeping this test set (example; modifiers-override)