fix(rslint_parser): Validate TS class member modifiers (#2192)
The different class members allow (some) of the following modifiers:

* `private`, `protected`, `public`
* `readonly`
* `override`
* `abstract`
* `static`
* `declare`

This PR adds the necessary violations for the TypeScript modifiers (all except `static`) so that the parser emits an error if:
* If a modifier appears out of order
* If a modifier can't be used with another modifier
* If a modifier can't be used for this member

It further adds support for parsing the `override` modifier.

This PR introduces the new `smallvec` dependency to store the parsed modifiers. This has two benefits compared to the previous data structure. 

* No heap allocations for all valid member modifiers (longest to my knowledge is `private declare readonly static member`
* Smaller footprint then allocating a fixed array with slots for all possible modifiers
* Centralized verification of the modifiers. It's no longer needed to first deduplicate the modifiers when parsing and then validating the combinations in a second pass when it's know what the member kind is. 

Part of #2187