refactor(rslint_parser): Lex contextual keywords as keywords (#2208)
This PR is part of #2035. The motivation for tackling this now is that I want to reduce the `Parser`s API. Currently, the parser exposes the following API:

* `cur`: Returns kind of current token
* `cur_src`: Returns the `string` representation of the current token
* `cur_tok`: Returns the `Range` of the current token.
* `nth`/`nth_src`/`nth_tok`: Same as the `cur` pendants but for the `nth` token (lookahead).

The `cur_src` and `cur_tok` methods are necessary for diagnostics (to get the range or print the actual token that was present in the source). However, `nth` should be limited to only allow tests on the `kind` because the only use should be to do lookahead. The main use of `nth_src` is the lookahead for contextual keywords. 

This PR changes the lexer to lex contextual keywords as keywords (for example, lex `yield` as `YIELD_KW` instead of `IDENT`) so that `nth_src` and later hopefully `nth_tok` can be removed. This involves some widespread changes because

* We now need to match on the keyword kind instead of `IDENT` in all match expressions
* Replace all usages of `eat_contextual` and `expect_contextual`
* The parser now needs to remap contextual keywords to identifiers (or names) when in places where they aren't keywords. 

This change has the nice side effect that we can now emit diagnostics if:
* a keyword uses any escape sequences (not allowed): `\u0061wait 0;`
* an identifier matches the text of a keyword in an identifier position and uses escape sequences (not allowed): `var x = ({ l\u0065t }) => {};`

Part of #2035