fix(rslint_parser): Fix ambiguity of the `<<` and `>>` tokens
The `<<` and `>>` (and `>>>`) can either be a left/right shift or the start/end of type arguments: `A<<B>(b: B) => number>>`.

This PR makes use of #2219 and re-scans these tokens to correctly distinguish them in the different contexts:

* `<<`: Re-lexed in places where a type argument is allowed to `<` `<`. Motivation, binary expressions are more frequent and re-lexing has a performance penalty because it evicts all lexing caches.
* `>` | `>>` | `>>>` | `>=` | `>>=` | `>>>=`: Re-lexed in the binary operator position.

I decided to not re-lex `>` in type argument end positions because it would require re-lexing to test if we're at a list end, and it changes way more tokens.