refactor(rslint_parser): Incremental lexing (#2219)
Implementation of a lazy-lexer that lexes the tokens when requested by the parser instead of lexing all tokens ahead of time. Part of #2035 

The motivation for this change is that it allows re-lexing a token in a different context. This is important in two scenarios:

* Error recovery: For example, the `/` token can either be the division operator in a binary expression or signal the start of a regular expression. The old implementation inferred whether it's the former or latter by internally tracking if it's a possible expression start and that worked decently enough for valid syntax. However, that wasn't the case if the parser performed error recovery which resulted in the lexer and parser get out of sync.
* When lexing `<<`. This can either be a left shift or in `B<<A>(a: A) => undefined>` a generic type where the type argument is an arrow function with type arguments. The current implementation has a `bump_multiple` operation so that the lexer can lex the `<` as individual tokens that then get grouped together to `<<` if the token appears in a binary operator position. The problem is that the parser operates on non-trivia tokens. So testing `p.at(T![<]) && p.nth_at(1, T![<])` yields `true` for `<<` but also `<   <` or worse, `< /* some comment */ <`. Both aren't valid left shift operators. That's why the parser must call back into the lexer that ensures that only `<<` without any trivia in between gets lexed as a left shift operator.


This PR refactors much of the Lexer, TokenSource and some of the Parser API:

### Lexer

The lexer now tracks the current token and calling `next_token` only returns the kind together with a potential diagnostic. Storing the current position is necessary because the lexer no longer uses an internal state to know if it's at the start of an expression to determine if a regular expression is allowed, and instead relies on the parser to call `re_lex` for regular expressions. However, that requires that the lexer knows the position of the current token and it's kind to determine if it should be re-lexed in the current context. 

I introduced some more helper methods in the Lexer and renamed a couple of methods. This helped me better understand what the Lexer does. 


### Buffered Lexer

Rome's parser heavily depends on the ability to look at the `nth` token and not just at the current token. The `BufferedLexer` adds a buffer in front of the `Lexer` that in the best case, ensures that every token is only lexed once. There are some situations where this isn't possible, for example if the parser calls into `re_lex`, performs a `rewind`, or lexes a token in another context. 

Without a buffered lexer, every `nth` call would need to take a snapshot of the lexer, parse all tokens up to the `nth` token, and then rewind the `Lexer` to its previous position. My first implementation did just that and it caused a 40% runtime regression. 


### Token Source

The `TokenSource` is similar to the `Lexer`, it returns tokens by tokens for the `Parser` to consume, but it skips `Trivia`. The old implementation operated on a list of lexed tokens. This is no longer the case because re-lexing can invalidate any following tokens. 

The new `TokenSource` directly calls into the `BufferedLexer` whenever it needs the next token. It uses an internal cache that stores the `nth` non-trivia token. This improves by roughly 3-5%. 

The `TokenSource` now further collects all processed `Trivia`. Collecting the `Trivia` is necessary for the lossless tree sink as well as for being able to compute the offset of each `Token`. 

The `TokenSource` no longer stores all `Token` because this information was mostly redundant to what the `Parser` already stores in the `Event`s. Only storing the `Trivia` without the `Token`s was another 3-5% performance improvement.

### Rewrite Parser

This is a change that I didn't want to make but was forced to. Up to now, `rewrite_events`, which is used to rewrite an expression tree to an assignment tree, operated directly on the `Parser`. This worked well because the `TokenSource` was immutable, the tokens weren't able to change (the token for a specific token index always returned the same kind and range). This assumption is no longer true if the parser ever calls into `re_lex` or lexes a token with a different context. A tokens' kind can then change, for example:

```
a[`b`] = "test"
   ^
```

Calling `p.nth(1)` returns `ident` if the parser's positioned at the `BACKTICK`. However, the kind of `p.cur()` is `TEMPLATE_CHUNK` after calling `lexer.next_token(LexContext::Template)` (moving the lexer ahead). 

That's why rewriting would need to keep track of when the parser called into `re_lex` or `next_token` with a context other than `Regular` OR we detach the `Parser` during rewriting from the token source and treat it as being immutable (it isn't allowed to add, change, or drop any tokens during a rewrite). That's what the `RewriteParser` is for. 


### Other API improvements

* `p.expect_keyword` and p.eat_keyword` are no longer necessary. The parser can now use `self.tokens.has_unicode_escape()` to determine if the current identifier contains any unicode escape and then emit a diagnostic if the current kind is a keyword. 


### Illustration

[Small illustration](https://excalidraw.com/#json=w3k_armSTSgK3socnG2cP,aH7QT4vgbT7HVzVxR1aUqw) explaining the differences before/after.