feat(rslint_parser): Parse TS-Decorators into Trivia (#2253)
This PR adds parser support for TypeScript's decorators by parsing the decorators into a new `skipped` token trivia. The reason for parsing it as trivia rather than just bumping the tokens is that a `skipped` trivia doesn't add an unexpected token to the parent node. Adding the unexpected node would make the parent node an `*UNKNOWN` node with the result that the formatter bails out of formatting the note completely. 

The implemented parsing doesn't perform any validation because the main goal is to be able to "skip" only decorators, kind of as an "intelligent" error recovery strategy.