fix(rome_formatter): Don't remove parenthesis from member assignments
## Summary

I played around with our formatter and noticed that formatting the benchmark test files twice results in the following parse errors:

```
error[SyntaxError]: Invalid assignment to `s || (s = Object.create(null))[i]`
     ┌─ vue.global.prod.js:4372:44
     │
4372 │                             this._props[i] = Z(this._props[i]), s || (s = Object.create(null))[i] =
     │                                                                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ This expression cannot be assigned to
```

The issue is that the formatter removes the parenthesis around the `object` of a static or computed member assignment.

This PR ensures that parenthesis around removed from the left-hand sie of a member assignment which is in line with what the formatter does for expressions.