fix(rome_js_parser: JSX `>=` parsing
## Summary

The parser failed to parse `<span>=</span>` because the lexer lexed `>=` as a single token rather than a `>` and a `=` token.

This PR changes the lexer to always lex `>=` (and `>>=` and `>>>=` as single tokens and only lex them as binary operators when called with `ReLex::BinaryOperator`.

## Test Plan

Added a new test verifying that `>=` works correctly