fix(rome_js_parser): Fix directive parsing (#2323)
## Summary

The parser incorrectly parsed any expression starting with a string literal as a directive. For example, it parsed the following expressions as directives that it shouldn't

```js
"use directive".length;
"use directive"[0];
"use directive"();
...
```

This PR changes the way we parse directives by now calling into `parse_expression` and then changing the expression to a directive only if it is a `StringLiteralExpression`.

The main challenge of doing this is that the expression either must be parsed into the directive or statement list. Luckily, the parser already has the infrastructure to achieve this by using a combination of `undo` and completing a marker again.

I then noticed that the formatter isn't normalizing the quotes for all string literals. I added a new `format_string_literal_token` helper and made sure that it is called for every string literal in the grammar. That's why there are so many formatter changes.

## Tests
Added a new parser test