fix(rome_js_parser): Fix TS Decorator with arrow param parsing
#2327 changed the parser to not store the last parsed token but instead
retrieve it by getting the kind from the last `Token` event in the parser events.

This didn't play well with skipping tokens because the parser didn't add any tokens to the `events` collection while in `skipping` mode.

This PR changes the `skipping` mode so that it only changes whatever the parser calls `bump` or `skip_as_token_trivia` but the parser will always add it to the `events` collection.

This adds a few unnecessary writes but this should be neglectable because

* Skipping token is rare and even then, often limited to very few tokens
* Moving the `bump` out of the condition may even help with branch prediction.

Fixes #2358