fix(rome_js_parser): Regex assignment (#2382)
An assignment to a regex (or any other token that uses re-lexing) may result in a panic if it is the last expression in the file. The source of the problem is that the code wrapping an assignment in an unknown assignment target rewinded the parser and then consumed all tokens up to the same position. However, this results in re-lexing all tokens in "default" mode.

This PR fixes the issue by simply changing the kind of the completed node to `UNKOWN_ASSIGNMENT` rather than using some custom recovery code. The result is more in line with our other error recovery logic that tries to retain valid sub-trees whenever possible.