fix(rome_rowan): Trivia-Piece text perf (#2381)
The existing `SyntaxTriviaPiece.text()` method called into `SyntaxTrivia.text_range()` which is expensive because it must iterate over all pieces to get the absolute offset.

The result was that `fourslash/reallyLargeFile` took minutes on my computer to format.

This PR changes the `piece.text` implementation to instead retrieve the text directly from the token because the piece is already aware of its absolute offset in the file, making this a very cheap operation

There are still some places where the implementation calls into `text_range` to get the offsets which always requires iterating over all trivia pieces to get the total length. This happens when calling the `pieces` iterator. We may want to consider caching the total length on the `GreenTrivia`.

## Tests

Formatting all TS files in the coverage directory now takes less than 2 seconds.