fix(rome_js_parser): Assignment in decorator (#2386)
Assignments are hard...

This PR fixes an issue where an assignment is used inside a decorator, e.g. `@test(--a)`

The parser used to panic when rewriting the assignment because any skipped tokens (the case for all tokens inside a decorator) are in the trivia collection and the parser events to ensure all parser methods work correctly.

However, this had the effect that `skip_trivia` in `rewrite_parser` moved the offset passed all skipped tokens, passed any not yet processed token (that ultimately get skipped).

This PR makes the rewrite parser aware of skipped tokens and only consumes them when the corresponding token gets bumped.