chore: Print lines exceeding line widths
Add a new section to formatter tests that print the lines that exceed the configured max line width.

There are few cases where it's valid that a line exceeds the configured line width, for example, if a single identifier exceeds that length. However, for most cases, it's a formatting error.