refactor(rome_rowan): Use associated types for `Ast*` traits
The `AstNode`, `AstNodeList`, and `AstSeparatedList` currently use generics to
parametrize the `Language` and `AstNode` (in case of the lists).

This PR changes the `Ast*` traits to use associated types instead. The semantic
difference of this change is that a single struct can no longer implement the traits for different languages.

Or, as explained in the Rust book:

> With associated types, we don’t need to annotate types because we can’t implement a trait on a type multiple times. In Listing 19-12 with the definition that uses associated types, we can only choose what the type of Item will be once, because there can only be one impl Iterator for Counter. We don’t have to specify that we want an iterator of u32 values everywhere that we call next on Counter. [source](https://doc.rust-lang.org/book/ch19-03-advanced-traits.html)

For example, the following will no longer be possible:

```rust
impl AstNode<JsLanguage> for JsAnyExpression {... }
impl AstNode<CssLanguage> for JsAnyExpression { ... }
```

I don't see a use case where we want to implement `AstNode` for different languages for a specific node.
The only use case that I'm aware of that gets close are super languages, but I think this should instead use a different `Language` and be union types instead.

 Using associated types allows us to create generic trait implementations for `AstNodeList` that otherwise aren't possible because a single node could implement the same trait multiple times:

 ```rust
 impl<T> MyTrait for T
   where T: AstNodeList {
 }
 ```