refactor(rome_js_formatter): Track token offsets instead of tokens (#2481)
The `Formatter` tracks all formatted token in debug builds to catch if the same token gets formatted twice or not formatted at all.

The current implementation tracks the `JsSyntaxToken`s. The downside of doing this is that a language agnostic implementation needs to track `SyntaxToken<L>` which:

* has the consequence that `Formatter` must be generic over `L`
* which in turn has the consequence that `Format` must be generic over `L`

This PR changes the node tracking to instead only track the absolute offsets of the formatted nodes rather than holding on to the node itself. This removes the need for a generic `Language` parameter and provides the same functionality.

## Test

* Commented out the `track_replaced` call in `format_replaced` and verified that some tests now panic with a message that token X wasn't formatted.
* Added one an additional `track_replaced` call to `format_replaced` and verified that tests are now failing with a message saying that the token X was formatted twice.