fix(formatter): Stack overflow in formatting computed member expressions
## Summary

The formatting stack overflowed in release/printed tokens twice for computed member expressions that have a computed member expression member.

The problem was that the upwards traversal logic didn't check if the `current` node is indeed the `parents.object()` node which is the rule it uses when traversing downwards.

## Test

I added a spec test and verified that it fails without the fix.