refactor(Formatter): `formatted` Macro (#2557)
This PR introduces a new `formatted` macro to format objects that are either `FormatElement`s or implement `Format`. The macro is needed to implement the shared formatter infrastructure #2525 without regressing our API. 

The proposed design in #2525 is that each formatter crate defines its own `AsFormat` and `IntoFormat` trait. Objects implementing either of these traits don't implement `Format` themselves (because of Rust's orphan rule) but expose a `format` and `into_format` methods that return an object that implements the formatting for it. 

The main challenge with this design is that `AstNode.format` doesn't return `FormatResult<FormatElement>` but an object implementing `Format` and `format_elements!` doesn't support combining `FormatElement`s with objects implementing `Format`. 

I considered changing the `AsFormat` trait to have two methods: `as_format` and `format` where `format` has a default implementation that calls `as_format().format(formatter)`. This worked surprisingly well except for the `FormatOptionalTrait` because that works on object implementing `Format`. Formatting an optional object that implements `AsFormat` would then have required to call `if_stmt.else_clause.as_format().format_or_empty(formatter)` which felt excessive. 

That's why this PR introduces a new `formatted!` macro similar to Rust's `write!` macro. The macro accepts a `Formatter` followed by the elements that should be formatted. It's result is `FormatResult<FormatElement>`. It accepts any object that implements the new `IntoFormatElement` trait. `IntoFormatElement` is similar to `Format`, except that it takes ownership of `self`. This new trait is necessary because implementing `Format` for `FormatElement` can't be done in an efficient way because it would be necessary to `clone` the `FormatElement` with each `format` call. 

The introduction of the `formatted` macro starts a shift of our formatter design from a `FormatElement` focused formatting to a logic where the formatting logic centers around the `Format` macro. I believe this will be beneficial over time and address some of the seams that have become visible in our current formatter design where building reusable logic around formatting a sequence of `Format` object felt awkward, because it often time required iterating over them first to collect any potential error that happens when formatting the object, and then returning an iterator over the `FormatElement`s. 

The final design that I've in mind is that our formatter changes so that even `concat`, `group`, `join` accept objects that are formattable AND the IR builders (`space_token`) don't return a `FormatElement` but an object that implements `Format` too. This will also bring us closer to a design where it will ultimately become possible to write to a shared buffer rather than having each `Format` impl return a `FormatElement`. `FormatElement` really is just the IR relevant for printing but should be less important when implementing the formatting logic. 


The main downside of this implementation is that there are now two macros and engineers need to know when to use `format_elements` and when `formatted`. This may even extend to three macros in the future, a `format_args` macro if we decide to implement `space` etc. to return `Format` objects.