refactor(formatter): Extract language specific options
This PR extracts the language specific `quote_style` option out of `rome_formatter` and moves it into `rome_js_formatter`.

This is done by adding a new associated type `Options` to `Format` that specifies the options passed to the formatter. I also changed the `Formatter` type to have one generic `Options` parameter.

This PR introduces a new `FormatOptions` trait that must be implemented by all option types that are used in conjunction with any `format` function.