refactor: Printer "fits" definition
This PR redefines how the printer decides if some content fits on a line or not.

Up to now, the printer only tested if the content (and the nested groups) fit on the line, but not if any content coming after that group fits too.

For example, if you have the following statement:

```js
if (someLongCondition) {}
```

where `(someLongCondition)` is grouped. The current printer considered that the content fits if printing the `)` didn't exceed the line width.

The new behavour changes the behaviour so that it tries to print any remaining content until it finds the first `\n` or exceeds the line width. This means for the example above that it only considers the content to fit if the `{}` also fit on the line without exceeding the line width.

This change has wide reaching consequences and regresses our formatting in many places because

* Our implementation often wraps content where it shouldn't
* And misses grouping in many places where it should

I don't intend to fix these issues as part of this PR, instead, this should be approached incrementelly by re-visiting every formatting rule and making sure that Rome creates the same IR as prettier.

Landing this change is important because it significantly changes how the printing works and the new behaviour is necessary to be able to adopt an IR output that's closer/similar to Prettier's.