feature(formatter): Add `BestFitting` IR element (#2591)
This PR introduces the new IR element `BestFitting`. The IR matches Prettier's `ConditionalGroupContent`. This IR can be useful if the best representation depends on the available space. `BestFitting` defers the choice of how to print the element to the printer by providing multiple variants. The printer picks the element that best fits given the circumstances (that's where its name is coming from but I'm open for better names).

The printer picks the first variant that fits on the current line and falls back to print the last variant in expanded mode if no variant fits.