refactor: Stabilize formatter (#2597)
Our JS Formatter is over-fitting on the test suite and it's very easy to introduce some stability issue when making any change to the formatter/printer. 

I was able to narrow down the main sources of the instability issues and fixed them as best I could. However, this PR introduces two regressions that I want to address in follow up PRs.

1. Trailing commas in call arguments: call-arguments now always print a trailing comma. Fixing this requires more extensive rules that use `BestFitting`. See #2421 for the follow-up PR. 
2. Formatting of inline comments. Inline comments aren't moved to the end anymore. This isn't ideal, I'm working on a follow up PR that re-defines the comment printing 

This PR removes `HardGroup`s. The main purpose of `HardGroup`s was to more deterministically handle inline comments. This is no longer needed because this PR removes this inline comment printing and we need to come up with a comment printing that doesn't require every node to be wrapped inside of a `HardGroup`. 

The "non-break" behaviour can be achieved by either just creating a sequence of tokens separated by spaces or using `BestFitting` where needed.