fix(parser): Fix JSX parser dropping content
Reverts #2508 because `parse_any_jsx_tag` returns `Absent` even when it parsed out the opening element and the tags content.

PR #2508 added a checkpoint in the parse binary expression to rewind if that happens but not to any other call sites of the function. Parsing a tag with an incomplete JSX child then gets skipped by the parser because the parse list assumes that it reached the end.

PR #2508 mentions that it fixes a bug when parsing `let a = <number> undefined`. I haven't been able to reproduce this issue except if JSX is enabled but type assertions aren't valid in TS when JSX is enabled.