feat(formatter): Treat skipped token trivia as syntax error
The parser uses skipped token trivia to recover from unrecognised syntax. For example, it could skip over the token `5` in `import 5 from "test";` because that allows to continue parsing the rest of the `from` statement.

Up to this point, we decided to format nodes with skipped token as if the nodes don't contain any syntax error because skipped token is, so far, only used to skip over decorator syntax. The issue with that is that our parser doesn't guarantee that the skipped decorators actually are valid and don't contain any syntax errors AND skipped token trivia might be used for real syntax errors at any point in the future. Thus, we shouldn't rely on that skipped token trivia is only used for syntax that Rome doesn't fully support.

This PR renames `FormatError::MissingRequiredChild` to `SyntaxError` and changes the leading trivia formatting to return `Err(FormatError::SyntaxError)` if any piece is a skipped token trivia to align it with how the rest of the formatter handles syntax errors.

This results in a few regressions on decorator formatting. The fix for that is to implement decorator parsing support.