refactor(rome_formatter): Add `Interned` `FormatElement`
Introduces a new reference-counted `FormatElement`. Useful for situations where the same content is part of the resulting IR twice but with different parent formatting. To this point, it has been necessary to clone the whole sub-tree. The ref counted FormatElement on the other hand only requires bumping the ref count.

`cargo test`