fix(rome_js_formatter): Fix suppression command handling for `if-else` clause
The suppression comments are checked as part of a node's formatting. It is, therefore, required that the formatter calls `format` on every node or the formatter won't respect the suppression comment.

This hasn't been the case for else clauses where the `if` statement formatting also formatted its else clauses.

This PR changes this by moving the logic for formatting the `else` clause into the `else_clause` formatting and calling it from the `if` stmt formatting.

This PR further ensures that this requirement upholds for static and computed members as well as sequence expressions. It isn't strictly required for these (yet) because the suppression comment will be associated with the inner expression of that node but it would become important if comments get attached to nodes.