refactor(rome_formatter): Implement `CommentStyle` on `FormatContext`
This is a small refactor that implements `CommentStyle` on `FormatContext` to avoid the additional `S` type parameter for all token builders.

This PR further simplifies the spacing handling between comments and tokens by fully delegating the spacing handling from the comments formatting to when a token gets formatted. This simplifies the overall logic.