refactor(rome_js_formatter): Remove unnecessary groups from `format_separated`
`format_separated` used to group all elements by default. This is often undesired and also comes with a performance cost (printer needs to test if the elements fit).

This PR changes the default of `format_separated` so that it doesn't group the elements except if enabled explicitly using `nodes_grouped`.

This PR further fixes the formatting of named import specifiers to more closely match Prettier's foramtting.