refactor(rome_formatter): Pre-compute suppressions
This PR changes how the formatter extracts formatter suppression comments. The existing implementation used to traverse to a node's first token and then checking if it has any suppression comment when `context.is_suppressed()` was called.

This PR iterates over the CST before starting the actual formatting to extract all suppressed nodes and stores them in an index that can be queried during formatting.

The benefit of pre-computing the suppressions is that the logic can then be combined with the logic that extracts a nodes leading/trailing comments because that also requires iterating over all trivia pieces in the tree.

Adding this traversal on its own would regress performance by about 10% which we can offset if combining it with the suppression handling because the use of `first_token` of the existing implementation is expensive.

This change has the added benefit that suppressions are now a core formatter concept.

### Other changes
This PR further fixes a few places where the formatter didn't correctly check if a node has any suppression comment and if so, prevents it from changing the formatting.

### Why add suppressions to the `FormatContext`

I decided to introduce a new `FormatNodeContext` that replaces the existing `CommentsContext` and `CommentStyle` and now also stores the `suppressions`.

The main reason for storing the suppressions on the context rather than on the `state` is that the context, unlike the `FormatState`, is language specific, and thus, it's possible to have a `Language` type parameter on the `Suppressions` type.

The one thing I dislike about adding it to the context is that the context is changed AFTER its creation.

### Performance

The performance of the two approaches are roughly the same.

### Future improvement
It would be possible to improve Rome's overall performance if the workspace provides a way to combine the visitors run for the linter and the visitor for computing the suppression comments.